WebStorage/tests/submissions/Infraware/storage_local_clear.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_local_clear</h3>
	</p>

	<hr>

	<div id="log"></div>
	
	<script>

		if (('localStorage' in window) && window.localStorage !== null){
			localStorage.clear();
			localStorage.setItem("name", "user1");

			if (localStorage.getItem("name") != null && localStorage.length == 1)
			{
				try {
					localStorage.clear();
				} catch (e) {
					test(function()
					{
						assert_unreached("clear method is failed.");
					}, "clear method test")
				}

				test(function() {
					assert_equals(localStorage.getItem("name"), null, "localStorage.getItem('name')")
				}, "result test of clear method (item)")

				test(function() {
					assert_equals(localStorage.length, 0, "localStorage.length")
				}, "result test of clear method (length)")

			} else {
				test(function() {
					assert_unreached("'name' item is not exist.");
				}, "item existence test")
			}
		} else {
			test(function()
			{
				assert_unreached("localStorage is not supported.");
			}, "localStorage support test")
		}

	</script>

 </body>
</html>
